Regression fix: wrong cross_prefix string
In commit f157632806fec7f96fe1bf97b6bac02d21fea1d7 , the x86_64 cross prefix is certainly wrong as it leads to /bin/bin.
For the i686 one, I'm not certain, but somehow not working for me. So to play safe, I'm proposing to revert that two lines to the old working one.