fix lsw-Vapoursynth output executable name
add target-os=mingw to fix the naming. Currently it gives a strange *.so file.
In addition, the build_ffmpeg line is a complete waste of build time. It is only useful in edge case when someone did not build ffmpeg but somehow directly called build_lsw.
Static is not an issue because the zeranoe build script by default... is static only
The issue reported for the vapoursynth plugin not building is due to QuickSync.

So either skip that unconditional ffmpeg rebuild, or check for ffmpeg libs before deciding rebuild or not.